I combined our models into model.py, added login and registr htmls and created a login base
