+++ webapp/__init__.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template
-
-from webapp.article_model import db, Articles
+from webapp.forms import LoginForm
+from gaming_web.webapp.model import db, Articles
 
 
 def create_app():
@@ -13,6 +13,12 @@ def create_app():
         title = 'Новости игр'
         news_list = Articles.query.order_by(Articles.written.desc()).all()
         return render_template('index.html', page_title = title, news_list = news_list)
+    
+    @app.route('/login')
+    def login():
+        title = 'Авторизация'
+        login_form = LoginForm()
+        return render_template('login.html', page_title=title, form=login_form)
     
     with app.app_context():
         db.create_all()
